minion: Fix machines being reported as disconnected
In order to check whether or not they were configured, the minion
performed a DB transaction to get its current config. Now the minion
saves it current config when it is set to avoid this transact when
the config is requested. This avoids the minions being listed as
disconnected when the DB transaction blocks.