db: Increase db lock granularity to the table level
Previously, if some goroutine wanted to perform a db transaction, it
would lock the whole database while doing so. But, almost all of our
transactions use only a few tables. This patch adds additional
functionality to the db, specifically a transaction object.
Transaction objects can only access a subset of tables, and lock
only those tables when performing transactions. This will allow
transactions that are independent to run concurrently in the
database, likely adding some small speedups and avoiding unneeded
blocking.

As a side effect, this patch lays some basic groundwork for
potential future patches to increase the concurrency of the db
further. Transaction objects have the ability to store state of
transactions, and could open up possibilites for things like
RW locks on the db, should the need arise.