db: Change Database API to support Conn-less access
Before, any function that wanted to access the database had to be
passed a Conn in order to do so. This leads to a lot of Conn objects
being passed around. Since we only ever use one database, it would
be cleaner to be able to open a connection to a default global
database from anywhere.

An advantage of this is that the code has less clutter; there are no
longer Conn's being passed as arguments simply to give other
functions access to the database. In addition, now that any function
can open up a DB connection, these connections can be restricted
better, since each function knows exactly what tables it needs to
access.

One consequence of this is that tests need to be written with more
care. Any test that involves a function that calls db.Open(...)
cannot be run in parallel and must call db.Reset() before calling
said function.