minion: Check that minions are running the same quilt as the daemon
If the minions and the daemon are not running the same version of
quilt, strange errors can occur, such as stitch JSON failing to parse
remotely but working fine locally. To mitigate this, the daemon now
sends its version number to the minions. If the versions aren't
compatible, the request will fail and an error reported.